refactor: rename modules
to be fair, name "modules" is not very descriptive. it's not the same as
"utils", but doesn't give any information about the content either. as i
understood, almost any nix file is a module, so "modules" is just a
directory which collects the files. that's not very helpful.

using my own username it's kind of funny, but more descriptive. it gives
a vibe "there is something custom here"; and it is. also, this way it
matches the "path" of options in the flake.

as a variant aside from my own username there was "default", and maybe i
will use default inside the flake, but not as the option "path".
"default.kitty.enable" - ok bro? i think using the username is better.

idea with the username is taken from etu/nixconfig on github. very
thanks